+++ internal/middleware/request.go
@@ -1,6 +1,8 @@
 package middleware
 
 import (
+	"time"
+
 	"github.com/labstack/echo/v4"
 	"go.uber.org/zap"
 )
@@ -11,7 +13,9 @@ func RequestLogger(logger *zap.Logger) echo.MiddlewareFunc {
 			req := c.Request()
 			res := c.Response()
 
+			start := time.Now()
 			err := next(c)
+			elapsed := time.Since(start)
 
 			logger.Info("Request",
 				zap.String("Remote IP", c.RealIP()),
@@ -21,6 +25,7 @@ func RequestLogger(logger *zap.Logger) echo.MiddlewareFunc {
 				zap.Int("Status", res.Status),
 				zap.Int64("Size", res.Size),
 				zap.String("User Agent", req.UserAgent()),
+				zap.Duration("Elapsed Time", elapsed),
 			)
 
 			return err
