--- pkg/util/slice.go
@@ -1,13 +1,0 @@
-package util
-
-func Filter[T any](slice []T, condition func(T) bool) []T {
-	var result []T
-
-	for _, v := range slice {
-		if condition(v) {
-			result = append(result, v)
-		}
-	}
-
-	return result 
-}
